【战斗】本地战斗 远端战斗 通用
【战斗】本地战斗 远端战斗 通用

+++ Assets/Script/Battle_Client/BattleShow/BattleClientDataTransfer/BattleClientDataTransfer.cs
@@ -71,13 +71,13 @@ namespace Battle_Client
     public class BattleClientMsg_BattleAttr
     {
         public EntityAttrType type;
-        public int value;
+        public float value;
     }
 
     public class BattleClientMsg_BattleValue
     {
         public EntityCurrValueType type;
-        public int value;
+        public float value;
     }
